Move `create_webhook_request` to a different module
