+++ src/tests/mod.rs
@@ -4,3 +4,4 @@ pub(crate) mod github;
 pub(crate) mod io;
 pub(crate) mod mocks;
 pub(crate) mod state;
+mod webhook;
